Makefile.toml: Clean up coverage commands (#999)
## Description

This commit cleans up the coverage commands by replacing the `test`
command logic with the `coverage-collect` logic and having all coverage
report generation rely on the `test` command instead. This allows us to
remove `test` from the `all` command, which was previsouly building and
running host based unit tests twice.

Additionally the commands related to failing when under a certain
coverage percentage have been removed as we can now soley rely on
codecov during CI for this.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

N/A

## Integration Instructions

N/A