Fix a potential race condition in EventDb/Allocator. (#1065)
EventDb/Allocator: fix a circular lock dependency where gcd_map_change()
callback could signal an event while downstream of an event lock.

This PR also removes the EVENT_DB_INITIALIZED static which should not be
required and masks this issue in test.

## Description

In some scenarios, an gcd_map_change() could occur as a result of an
allocation expansion while the event lock was held. Since signalling an
event required holding the event lock, this would cause re-entrancy and
deadlock.

An example of such a deadlock is when a runtime event (such as virtual
address change) is created, it goes into `runtime::add_runtime_event()`
which then allocates the event structure in the runtime allocator and
pushes it on a linked list. This is a pool allocation; but if that pool
allocation causes an expand in the runtime allocator, this will then
trigger a GCD change which will then signal an event for memory map
change resulting in a re-entrant deadlock of this nature. This can be
seen in unit testing if only the first commit of this PR is taken.

This change adds a mitigation for this by allowing events to be signaled
even if the event lock is held - if a signal is attempted while the
event lock is held, it is instead placed in a pending event queue; once
the event lock is released, then any pending events signaled while the
lock was held are drained (in signal order) into the event database.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Booted with aarch64 platform hardware with these changes.

## Integration Instructions

N/A