Add compile-time component parameter validation (#1124)
## Description

Today, component parameters are validated as follows:

- In debug build with debug asserts
- In test builds with debug asserts
- Not in release builds

This change both simplifies overall component definition to make their
construction more consistent and easier to follow in the codebase while
also introducing compile-time parameter validation for Patina-defined
parameter types.

Patina-defined parameter (`Param`) types include types like:

- `Config<T>`
- `ConfigMut<T>`
- `Hob<T>`
- `Service<T>`
- etc.

This allows the most common errors made using the pre-defined parameter
types to be detected very quickly during build. Custom parameter types
will be validated at runtime. A follow up change will prevent execution
of components with invalid custom parameter combinations and log an
error message alongside the table of components not dispatched.

---

The [trybuild](https://crates.io/crates/trybuild) crate is used in
`dev-dependencies`. It provides a clean way to test that invalid
parameter combinations have the expected compilation error messages. The
tests are placed in a `ui` directory following convention described in
the `trybuild` docs that these tests are focused on a user's interaction
with the code.

---

The main changes are:

- Entry points are now always called `entry_point` and do not require
any special annotation
- "Standalone" (free standing) function entry points are no longer
allowed and entry points must always be an `entry_point` method in an
`impl` block for the type
- Only a single `#[component]` attribute is needed on an `impl` block
for a component.

The is intended to make components easier to understand and review while
taking advantage of the changes to provide quick compile-time feedback
on component construction.

---

The `#[component]` attribute is applied onto the `impl` block
specifically because this allows visibility into the `impl` block to
check if an `entry_point` method is present which is a requirement.
Similarly, it provides visibility for any potential future component
special methods to be validated.

- [x] Impacts functionality?
- [ ] Impacts security?
- [x] Breaking change?
- [x] Includes tests?
- [x] Includes documentation?

## How This Was Tested

- `cargo make all`
- QEMU Q35 & SBSA boot to EFI shell

## Integration Instructions

- See the component documentation updates in the change

---------

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>