[AArch64] Move hardware interrupt handlers to be guarded by per-handler RwLock (#1165)
## Description

The current code guards all the interrupt handlers in a common TPL
mutex. This will have 2 side effects:

1. All interrupt handlers will run under the TPL defined by the handler
mutex, which is set to TPL_HIGH. This will prevent a lot of fundamental
functionalities from core to panic;
2. All interrupt IDs will be guarded by the same mutex, which could
potentially run into re-entrant mutex access if a different interrupt
comes in while the first interrupt is still being processed.

This change moves the hardware interrupt handler to be guarded using
per-handler RwLock to resolve above issues.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This change is tested on QEMU SBSA and physical platform.

## Integration Instructions

N/A

---------

Co-authored-by: Michael Kubacki <michael.kubacki@microsoft.com>