Rework memory_attributes_table to remove the MEMORY_ATTRIBUTES_TABLE static (#1173)
## Description

Remove `MEMORY_ATTRIBUTES_TABLE` global static - the pointer in question
is stored in the global configuration table. Instead of keeping a
duplicate copy, use the entry in the global table.
* adjust `core_install_configuration_table` return value to include the
old pointer if the table is removed or modified.
* add `get_configuration_table` API to return a pointer from the
configuration table for the given vendor guid.
* rework MAT logic to use the config table exclusively.
* Replace POST_RTB atomic bool with Once and add a `TestOnce` capability
for test support to allow resetting the Once for testing purposes.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Passes unit tests; booted on AARCH64 HW, checked MAT table as dumped in
UEFI logs.

## Integration Instructions

N/A