Replaced AtomicPtr with Once in boot_services and runtime_services (#1198)
## Description

Replaces the AtomicPtr usage with Once for semantics that better align
with intended usages. Also makes the `static` lifetime requirement on
BootServices (which was effectively there when it was being used as a
raw pointer) more explicit.

- [x] Impacts functionality?
- [ ] Impacts security?
- [x] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Passes unit testing, confirmed boot to EFI shell in QEMU Q35.

## Integration Instructions

Construction of `StandardBootServices` and `StandardRuntimeServices`
will need to supply a static-lifetime reference to the appropriate
structure. This should already be the case since most instantiations
will be from the global boot services or runtime services pointers from
the system table that already have `static` semantics (i.e. they live
for the lifetime of UEFI). For test usages, leaking a Box should
suffice.